requirements.txt
langchain_google_genai
langchain_core
python_dotenv
streamlit
langchain_community
chromadb
bs4
ipykernel
pypdf
langchainhub
langchain_chroma
sentence-transformers
langdetect
langid
openai-whisper
pyttsx3 
sounddevice
gTTS
pydub
playsound
edge_tts
torch
torchvision
torchaudio
google-generativeai
soundfile
SpeechRecognition
pyaudio
ttsvoice
chatterbox-tts
streamlit_webrtc
gradio
langchain_huggingface
parler
transformers
sentencepiece
ffmpeg-python
sacremoses